Connection class finished

+++ SDL_Pathfinding/SDL_Pathfinding/Connection.h
@@ -3,20 +3,18 @@
 #include <string>
 #include <map>
 
-//Totes les connexions seran bidireccionals
-
 class Connection
 {
 private:
 	std::string fromNodeId;
 	std::string toNodeId;
-	float cost;
+	int cost;
 
 public:
-	Connection(float _cost, std::string _fromNodeId, std::string _toNodeId);
+	Connection(int _cost, std::string _fromNodeId, std::string _toNodeId);
 	~Connection();
 
-	float GetCost() { return cost; }
+	int GetCost() { return cost; }
 	std::string GetFromNode() { return fromNodeId; }
 	std::string GetToNode() { return toNodeId; }
 };
